refactor(mcp): align module organization
