+++ apps/dashboard/src/lib/hooks/use-mcp-connection-toast.ts
@@ -2,17 +2,7 @@
 
 import { useEffect } from "react";
 import { toast } from "sonner";
-
-const MCP_OAUTH_ERROR_MESSAGES: Record<string, string> = {
-  mcp_oauth_denied: "MCP authorization was canceled.",
-  mcp_oauth_failed: "MCP authorization failed. Try connecting again.",
-  mcp_oauth_invalid_callback:
-    "The MCP authorization link is invalid or expired.",
-  mcp_oauth_refresh_token_required:
-    "This server did not provide a refresh token, so it cannot stay connected.",
-  mcp_oauth_session_required:
-    "Sign in again before connecting this MCP server.",
-};
+import { MCP_OAUTH_ERROR_MESSAGES } from "@/constants/mcp";
 
 export function useMcpConnectionToast() {
   useEffect(() => {
