+++ packages/ai/src/integrations/mcp.ts
@@ -8,6 +8,7 @@ import { customAlphabet } from "nanoid";
 import type {
   CreateMcpServerIntegrationParams,
   McpHeaderMap,
+  McpServerIntegrationSerializationInput,
   UpdateMcpServerIntegrationParams,
 } from "../types/integrations";
 import type { McpAuthType } from "../types/mcp-oauth";
@@ -43,31 +44,9 @@ async function assertOrganizationMember(
   }
 }
 
-export function serializeMcpServerIntegration<
-  T extends {
-    id: string;
-    name: string;
-    url: string;
-    description: string | null;
-    authType: string;
-    encryptedHeaders: McpHeaderMap;
-    enabled: boolean;
-    lastToolSyncAt?: Date | null;
-    toolSyncStatus?: string;
-    toolSyncError?: string | null;
-    indexedToolCount?: number;
-    oauthCredential?: {
-      status: string;
-    } | null;
-    createdAt: Date;
-    createdByUser?: {
-      id: string;
-      name: string;
-      email: string;
-      image: string | null;
-    } | null;
-  },
->(integration: T) {
+export function serializeMcpServerIntegration(
+  integration: McpServerIntegrationSerializationInput
+) {
   return {
     id: integration.id,
     name: integration.name,
