+++ packages/ai/src/utils/mcp-oauth-fetch.ts
